+++ src/app/contact/page.tsx
@@ -15,11 +15,11 @@ import {
   Textarea,
   Typography,
 } from "@mui/joy";
-import { EmailIcon, GitHubIcon, LinkedInIcon, SendIcon } from "@/_assets/icons";
+import { EmailIcon, GitHubIcon, LinkedInIcon, SendIcon } from "assets/icons";
 import Link from "next/link";
-import "@/_assets/stylesheets/contact.css";
-import useBreakpoint from "@/_hooks/useBreakpoints";
-import Hero from "@/_components/Hero/Hero";
+import "@/assets/stylesheets/contact.css";
+import useBreakpoint from "hooks/useBreakpoints";
+import Hero from "components/Hero/Hero";
 
 export default function Contact() {
   const lgScreen = useBreakpoint("lg");
