+++ tsconfig.json
@@ -20,10 +20,10 @@
     ],
     "baseUrl": "src/",
     "paths": {
-      "@/_components/*": ["_components/*"],
-      "@/_utils/*": ["_utils/*"],
-      "@/_hooks/*": ["_hooks/*"],
-      "@/_assets/*": ["_assets/*"]
+      "@/components/*": ["components/*"],
+      "@/utils/*": ["utils/*"],
+      "@/hooks/*": ["hooks/*"],
+      "@/assets/*": ["assets/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
